# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Guardian Cyber Safety & Threat Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRequest {
  /** The text to analyze */
  text: string;
}

export interface AnalysisResult {
  /** Classification category (Normal, Toxic, Scam, Threat) */
  category: string;
  /** Confidence score 0-1 */
  confidence: number;
  /** Risk level (Low, Medium, High) */
  riskLevel: string;
  /** Human-readable alert message */
  alertMessage: string;
  /** The text that was analyzed */
  analyzedText: string;
  /** The type of analyzer used */
  analyzerType: string;
  /** Record ID */
  id?: number;
  /** Timestamp of analysis */
  createdAt?: string;
}

export interface AnalysisRecord {
  id: number;
  category: string;
  confidence: number;
  /** Risk level (Low, Medium, High) */
  riskLevel: string;
  alertMessage: string;
  analyzedText: string;
  analyzerType: string;
  createdAt: string;
}

export type DashboardStatsRecentActivityItem = {
  hour: string;
  count: number;
};

export type DashboardStatsCategoryBreakdownItem = {
  category: string;
  count: number;
  percentage: number;
};

export interface DashboardStats {
  /** Total messages analyzed */
  totalAnalyzed: number;
  /** Count of toxic messages */
  toxicCount: number;
  /** Count of scam messages */
  scamCount: number;
  /** Count of threat messages */
  threatCount: number;
  /** Count of safe messages */
  safeCount: number;
  /** Recent analysis breakdown by hour */
  recentActivity: DashboardStatsRecentActivityItem[];
  /** Category distribution */
  categoryBreakdown: DashboardStatsCategoryBreakdownItem[];
}