# lib/api-zod/src/generated/types/analysisRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Guardian Cyber Safety & Threat Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRecord {
  id: number;
  category: string;
  confidence: number;
  /** Risk level (Low, Medium, High) */
  riskLevel: string;
  alertMessage: string;
  analyzedText: string;
  analyzerType: string;
  createdAt: string;
}