# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Guardian Cyber Safety & Threat Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  /** Classification category (Normal, Toxic, Scam, Threat) */
  category: string;
  /** Confidence score 0-1 */
  confidence: number;
  /** Risk level (Low, Medium, High) */
  riskLevel: string;
  /** Human-readable alert message */
  alertMessage: string;
  /** The text that was analyzed */
  analyzedText: string;
  /** The type of analyzer used */
  analyzerType: string;
  /** Record ID */
  id?: number;
  /** Timestamp of analysis */
  createdAt?: string;
}