# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Guardian Cyber Safety & Threat Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsCategoryBreakdownItem } from "./dashboardStatsCategoryBreakdownItem";
import type { DashboardStatsRecentActivityItem } from "./dashboardStatsRecentActivityItem";

export interface DashboardStats {
  /** Total messages analyzed */
  totalAnalyzed: number;
  /** Count of toxic messages */
  toxicCount: number;
  /** Count of scam messages */
  scamCount: number;
  /** Count of threat messages */
  threatCount: number;
  /** Count of safe messages */
  safeCount: number;
  /** Recent analysis breakdown by hour */
  recentActivity: DashboardStatsRecentActivityItem[];
  /** Category distribution */
  categoryBreakdown: DashboardStatsCategoryBreakdownItem[];
}